requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2
djangorestframework==3.12.4
image==1.5.33
numpy==1.20.2
opencv-python==4.5.1.48
passlib==1.7.4
Pillow==8.2.0
pycparser==2.20
pytesseract==0.3.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
